routes: export handlers only, so the webpack build's route type check passes
The webpack build (unlike Turbopack's) generates Next's route module
checks under .next/types, and six routes failed them: a handler whose
request was optional, and five route files exporting helpers and
constants beside their handlers. Each moves to where it belongs —
MAX_PUT_BODY_BYTES to coding-project-tree.ts, MAX_AUDIO_BYTES to
transcribe-limits.ts, the Hermes exit strings to hermes-chat-exit.ts
(the tests' door, in place of the route's __test object), the transcribe
model read from stt-preference.ts — or stops being exported.
